Limit op data at sector granularity (#101)
* feat(kitsune2_api): Remove limit_bytes from retrieve_op_hashes_in_time_slice

 It is better to limit between slices than within them.

* feat(kitsune2_core): Follow new API to not limit retrieve_op_hashes_in_time_slice

* feat(kitsune2_dht): Limit the number of bytes retrieved in a DHT diff in a less granular way. No longer by op but on natural boundaries

This could have resulted in a sync stalling because the provided limit was too low to fetch a full sector or slice. It's better to exceed the (soft) limit than risk stalling

* feat(kitsune2_gossip): Integrate DHT changes for soft limiting data fetches at natural boundaries

+++ crates/gossip/src/respond/disc_sector_details_diff_response.rs
@@ -52,7 +52,7 @@ impl K2Gossip {
             used_bytes,
             state.peer_max_op_data_bytes,
         );
-        state.peer_max_op_data_bytes -= used_bytes;
+        state.peer_max_op_data_bytes -= used_bytes as i32;
 
         match next_action {
             DhtSnapshotNextAction::CannotCompare
