+++ crates/gossip/src/respond/ring_sector_details_diff_response.rs
@@ -50,7 +50,7 @@ impl K2Gossip {
             .await?;
 
         if let Some(state) = state.as_mut() {
-            state.peer_max_op_data_bytes -= used_bytes;
+            state.peer_max_op_data_bytes -= used_bytes as i32;
         }
 
         match next_action {
